backend/requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
python-dotenv
pytesseract
pdfplumber
pillow
pdf2image
ollama
